Date: Thu, 13 Mar 2003 00:57:39 -0800 From: Devin Carraway <qpsmtpd-list@devin.com> To: qpsmtpd@perl.org Subject: HELO hook and check plugin
Speaking of direct-to-MX spam, both AOL and Yahoo are large companies
with whole walls-full of servers devoted to mail delivery.  None of them
announce themselves with "HELO yahoo.com" or "HELO aol.com."  Spammers
certainly do, though.

Here's a patch to SMTP.pm to add hooks for HELO and EHLO, and a plugin
to use them.


git-svn-id: https://svn.perl.org/qpsmtpd/trunk@119 958fd67b-6ff1-0310-b445-bb7760255be9