Call add_recipient correctly in kavscanner plugin
plugins/virus/kavscanner calls $transaction->add_recipient($_->address)
on a list of Mail::Address objects, but add_recipient() clearly
documents that it takes Qpsmtpd::Address (or compatible) objects, not
strings.  This is a bit of a drive-by fix inspired by a grep through the
codebase for calls to add_recipient().