spamassassin plugin tests and updates
parsing of the X-Spam-Status header now uses split instead of a regexp. Mining a collection of mail headers revealed that the regexp was not as robust as it could be. Now the parsing should work for everyone, regardless of how they configure SpamAssassin.

the spamassassin note is now a hashref (was : delimited fields). More data is now available, its more easily accessible (no unpacking required), and its extensible so it won't require future modification. I just added the spamassassin note feature so it's unlikely this change will hurt anyone.

Added logging for many more error and status conditions.

renamed reject_threshold option to 'reject', to be consistent with other plugins that have a 'reject' option.