headers: new plugin
Checks for missing or empty values in the From or Date headers.

Make sure no singular headers are duplicated. Singular headers are:

 Date From Sender Reply-To To Cc Bcc
 Message-Id In-Reply-To References Subject

Optionally test if the Date header is too many days in the past or
future. If future or past are not defined, they are not tested.

If the remote IP is whitelisted, header validation is skipped.