postfix: avoid logging full headers
When the postfix queue plugin was added in 5abf363 (Added Postfix queue
plugin thanks to Peter J Holzer!, 2004-02-19), some leftover debugging
seems to have been included.

The 'print STDERR "hdr: $_\n";' line causes the full headers of every
message to be logged.  A few lines below, a similar statement for the
body lines is commented out.  Do the same for the header statement.