Change Authentication-Results "me" file and expand
Per request, make the Authentication-Results server ID config file
"me-auth-results" for clarity.

Also, expand its meaning slightly - use "none" to disable adding or
modifying Authentication-Results headers.  This is useful when qpsmtpd
is used in an internal hop and should not be overriding an edge hop that
checked SPF/DKIM/etc.