Allow an alternate ID for Authentication-Results (#323)
* Allow an alternate ID for Authentication-Results

When using a cluster of servers, it's sometimes needed to have the same
ID in the Authentication-Results header, rather than just the hostname,
and you don't always want to change "me" (because that has other
effects).  Allow an alternate "ar-me" config file.

* Change Authentication-Results "me" file and expand

Per request, make the Authentication-Results server ID config file
"me-auth-results" for clarity.

Also, expand its meaning slightly - use "none" to disable adding or
modifying Authentication-Results headers.  This is useful when qpsmtpd
is used in an internal hop and should not be overriding an edge hop that
checked SPF/DKIM/etc.